Shop finished

+++ src/main/java/com/example/Uncha_Muncha_Bot/dto/ShopTypeDTO.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 public class ShopTypeDTO {
     private Long id;
-    private Integer shopId;
+    private Long shopId;
     private String shopType;
 }
